Remove serverless dependency on agent-payload/v5/process.
