+++ pkg/forwarder/telemetry.go
@@ -8,18 +8,12 @@ package forwarder
 import (
 	"expvar"
 
-	model "github.com/DataDog/agent-payload/v5/process"
-
 	"github.com/DataDog/datadog-agent/pkg/forwarder/endpoints"
 	"github.com/DataDog/datadog-agent/pkg/forwarder/transaction"
-	"github.com/DataDog/datadog-agent/pkg/orchestrator"
 	"github.com/DataDog/datadog-agent/pkg/telemetry"
-	"github.com/DataDog/datadog-agent/pkg/util/log"
 )
 
 var (
-	transactionsIntakeOrchestrator = map[model.K8SResource]*expvar.Int{}
-
 	highPriorityQueueFull            = expvar.Int{}
 	transactionsInputBytesByEndpoint = expvar.Map{}
 	transactionsInputCountByEndpoint = expvar.Map{}
@@ -74,23 +68,6 @@ func initEndpointExpvars() {
 	for _, endpoint := range endpoints {
 		transaction.TransactionsSuccessByEndpoint.Set(endpoint.Name, expvar.NewInt(endpoint.Name))
 	}
-}
-
-func initOrchestratorExpVars() {
-	for _, nodeType := range orchestrator.NodeTypes() {
-		transactionsIntakeOrchestrator[nodeType] = &expvar.Int{}
-		transaction.TransactionsExpvars.Set(nodeType.String(), transactionsIntakeOrchestrator[nodeType])
-	}
-	transaction.TransactionsExpvars.Set("OrchestratorManifest", &transactionsOrchestratorManifest)
-}
-
-func bumpOrchestratorPayload(nodeType int) {
-	e, ok := transactionsIntakeOrchestrator[model.K8SResource(nodeType)]
-	if !ok {
-		log.Errorf("Unknown NodeType %v, cannot bump expvar", nodeType)
-		return
-	}
-	e.Add(1)
 }
 
 func initTransactionsExpvars() {
