+++ pkg/forwarder/telemetry_serverless.go
@@ -1,0 +1,13 @@
+// Unless explicitly stated otherwise all files in this repository are licensed
+// under the Apache License Version 2.0.
+// This product includes software developed at Datadog (https://www.datadoghq.com/).
+// Copyright 2016-present Datadog, Inc.
+
+//go:build serverless
+// +build serverless
+
+package forwarder
+
+func initOrchestratorExpVars() {}
+
+func bumpOrchestratorPayload(nodeType int) {}
